telegram-sidecar/requirements.txt
telethon==1.38.0
asyncpg==0.29.0
python-dotenv==1.0.1
